+++ src/ncc/CLI/Management/ProjectMenu.php
@@ -24,10 +24,13 @@
 
     use Exception;
     use ncc\Enums\ProjectTemplates;
+    use ncc\Enums\Scopes;
+    use ncc\Managers\CredentialManager;
     use ncc\Managers\ProjectManager;
     use ncc\Objects\CliHelpSection;
     use ncc\Utilities\Console;
     use ncc\Utilities\Functions;
+    use ncc\Utilities\Resolver;
     use ncc\Utilities\Security;
 
     class ProjectMenu
@@ -37,6 +40,7 @@
          *
          * @param array $args
          * @return int
+         * @throws Exception
          */
         public static function start(array $args): int
         {
@@ -48,6 +52,11 @@
             if(isset($args['template']))
             {
                 return self::applyTemplate($args);
+            }
+
+            if(isset($args['install']))
+            {
+                return self::installProject($args);
             }
 
             return self::displayOptions();
@@ -113,6 +122,93 @@
 
             Console::out(sprintf('Project successfully created in \'%s\'', $project_manager->getProjectPath()));
             Console::out(sprintf('Modify the project configuration in \'%s\'', $project_manager->getProjectPath() . DIRECTORY_SEPARATOR . 'project.json'));
+            return 0;
+        }
+
+        /**
+         * Installs the project dependencies
+         *
+         * @param array $args
+         * @return int
+         * @throws Exception
+         */
+        private static function installProject(array $args): int
+        {
+            if(Resolver::resolveScope() !== Scopes::SYSTEM)
+            {
+                Console::outError('You cannot install packages in a user scope, please run this command as root', true, 1);
+                return 1;
+            }
+
+            if(isset($args['path']) || isset($args['p']))
+            {
+                $project_path = $args['path'] ?? $args['p'];
+            }
+            elseif(is_file(getcwd() . DIRECTORY_SEPARATOR . 'project.json'))
+            {
+                $project_path = getcwd();
+            }
+            else
+            {
+                Console::outError('Missing option: --path|-p, please specify the path to the project', true, 1);
+                return 1;
+            }
+
+            $authentication = $args['authentication'] ?? $args['a'] ?? null;
+            $authentication_entry = null;
+
+            if($authentication !== null)
+            {
+                $entry = (new CredentialManager())->getVault()?->getEntry($authentication);
+
+                if($entry->isEncrypted())
+                {
+                    $tries = 0;
+                    while(true)
+                    {
+                        if (!$entry->unlock(Console::passwordInput('Password/Secret: ')))
+                        {
+                            $tries++;
+                            if ($tries >= 3)
+                            {
+                                Console::outError('Too many failed attempts.', true, 1);
+                                return 1;
+                            }
+
+                            Console::outError(sprintf('Incorrect password/secret, %d attempts remaining', 3 - $tries));
+                        }
+                        else
+                        {
+                            Console::out('Authentication successful.');
+                            return 1;
+                        }
+                    }
+                }
+
+                $authentication_entry = $entry->getPassword();
+            }
+
+            try
+            {
+                $project_manager = new ProjectManager($project_path);
+            }
+            catch(Exception $e)
+            {
+                Console::outException('There was an error while trying to load the project', $e, 1);
+                return 1;
+            }
+
+            try
+            {
+                Console::out(sprintf('Installed %s packages', count($project_manager->installDependencies($authentication_entry))));
+            }
+            catch(Exception $e)
+            {
+                Console::outException('There was an error while trying to install the project dependencies', $e, 1);
+                return 1;
+            }
+
+            Console::out(sprintf('Project dependencies successfully installed in \'%s\'', $project_manager->getProjectPath()));
             return 0;
         }
 
@@ -183,6 +279,7 @@
                 new CliHelpSection(['help'], 'Displays this help menu about the value command'),
                 new CliHelpSection(['create', '--path|-p', '--name|-n', '--package|--pkg', '--ext'], 'Creates a new ncc project'),
                 new CliHelpSection(['template', '--path|-p', '--name|-n'], 'Applies a template to the project'),
+                new CliHelpSection(['install', '--path|-p', '--authentication|-a'], 'Installs the project dependencies'),
             ];
 
             $options_padding = Functions::detectParametersPadding($options) + 4;